Major refactor to use bundling and transpilation for an easy distribution
