--- babel.config.js
@@ -1,13 +1,0 @@
-// babel.config.js
-module.exports = {
-  presets: [
-    [
-      "@babel/preset-env",
-      {
-        targets: {
-          node: "current",
-        },
-      },
-    ],
-  ],
-};
