--- src/defaultValues.js
@@ -1,19 +1,0 @@
-exports.key = "C";
-exports.scale = "lydian";
-exports.tempo = "240";
-exports.notes = "8";
-exports.empty = "None";
-exports.keys = [
-  "A",
-  "Bb",
-  "B",
-  "C",
-  "Db",
-  "D",
-  "Eb",
-  "E",
-  "F",
-  "Gb",
-  "G",
-  "Ab",
-];
