+++ src/server/api/middlewares/errorHandler.js
@@ -1,0 +1,19 @@
+import { createError } from "http-errors";
+
+export default (app) => {
+    // catch 404 and forward to error handler
+    app.use(function (req, res, next) {
+        next(createError(404));
+    });
+
+    // error handler
+    app.use(function (err, req, res, next) {
+        // set locals, only providing error in development
+        res.locals.message = err.message;
+        res.locals.error = req.app.get("env") === "development" ? err : {};
+
+        // render the error page
+        res.status(err.status || 500);
+        res.render("error");
+    });
+}
