+++ src/server/app.js
@@ -1,14 +1,11 @@
-var createError = require("http-errors");
 var express = require("express");
 var path = require("path");
 var cookieParser = require("cookie-parser");
 var logger = require("morgan");
-
-var indexRouter = require("./routes/index");
+import setup from "./api";
 
 var app = express();
 
-// view engine setup
 app.set("views", path.join(__dirname, "../client/views"));
 app.set("view engine", "pug");
 
@@ -18,23 +15,7 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, "../client/public")));
 
-app.use("/", indexRouter);
-
-// catch 404 and forward to error handler
-app.use(function (req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get("env") === "development" ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render("error");
-});
+setup(app);
 
 const port = 8000;
 app.listen(port, () => console.log(`Example app listening on port ${port}!`));
