+++ src/server/defaultValues.js
@@ -1,0 +1,21 @@
+export default {
+  key: "C",
+  scale: "lydian",
+  tempo: "240",
+  notes: "8",
+  empty: "None",
+  keys: [
+    "A",
+    "Bb",
+    "B",
+    "C",
+    "Db",
+    "D",
+    "Eb",
+    "E",
+    "F",
+    "Gb",
+    "G",
+    "Ab",
+  ]
+}
